refactor: Move internal Raptor roles to "raptor.rangeraptor.internalapi"
This get rid of a circular dependency
