+++ src/main/java/org/opentripplanner/transit/raptor/package.md
@@ -141,7 +141,7 @@ but there is no implementation for this.
 The debugger instrument the existing code for the given stops determined by the
 [`DebugRequest`](api/request/DebugRequest.java). If no debug listeners exist, then no debugging code
 is injected or run; hence the performance overhead under normal execution is minimal. The main
-Raptor logic will post events to the [DebugHandler](rangeraptor/view/DebugHandler.java)
+Raptor logic will post events to the [DebugHandler](rangeraptor/internalapi/DebugHandler.java)
 interface. There are one handler implementation for each event type(stop arrival, pattern ride, and
 path), all created by the [DebugHandlerFactory](rangeraptor/debug/DebugHandlerFactory.java). The
 handler implementations are called _Adapters_ because they take the internal Raptor event and
