+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/standard/besttimes/BestTimes.java
@@ -3,7 +3,7 @@ package org.opentripplanner.transit.raptor.rangeraptor.standard.besttimes;
 import static org.opentripplanner.transit.raptor.util.IntUtils.intArray;
 
 import java.util.BitSet;
-import org.opentripplanner.transit.raptor.rangeraptor.WorkerLifeCycle;
+import org.opentripplanner.transit.raptor.rangeraptor.internalapi.WorkerLifeCycle;
 import org.opentripplanner.transit.raptor.rangeraptor.transit.TransitCalculator;
 import org.opentripplanner.transit.raptor.util.BitSetIterator;
 import org.opentripplanner.util.lang.ToStringBuilder;
