+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/standard/stoparrivals/StopArrivals.java
@@ -3,7 +3,7 @@ package org.opentripplanner.transit.raptor.rangeraptor.standard.stoparrivals;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
 import org.opentripplanner.transit.raptor.api.transit.TransitArrival;
-import org.opentripplanner.transit.raptor.rangeraptor.RoundProvider;
+import org.opentripplanner.transit.raptor.rangeraptor.internalapi.RoundProvider;
 import org.opentripplanner.transit.raptor.rangeraptor.standard.BestNumberOfTransfers;
 import org.opentripplanner.transit.raptor.rangeraptor.standard.DestinationArrivalListener;
 import org.opentripplanner.transit.raptor.rangeraptor.transit.EgressPaths;
