+++ src/test/java/org/opentripplanner/transit/raptor/rangeraptor/transit/SlackProviderAdapterTest.java
@@ -6,7 +6,7 @@ import static org.opentripplanner.transit.raptor.api.transit.RaptorSlackProvider
 import org.junit.jupiter.api.Test;
 import org.opentripplanner.transit.raptor._data.RaptorTestConstants;
 import org.opentripplanner.transit.raptor.api.transit.RaptorSlackProvider;
-import org.opentripplanner.transit.raptor.rangeraptor.SlackProvider;
+import org.opentripplanner.transit.raptor.rangeraptor.internalapi.SlackProvider;
 import org.opentripplanner.transit.raptor.rangeraptor.lifecycle.LifeCycleEventPublisher;
 import org.opentripplanner.transit.raptor.rangeraptor.lifecycle.LifeCycleSubscriptions;
